chore(team): split memory flush helpers
